backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
ultralytics>=8.0.0
transformers>=4.36.0
torch>=2.1.0
torchvision>=0.16.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
